Added the following additional cost models to tiles domain: heavy, sqrt, inverse, reverse. Default is still unit. Specify cost model by using -cost [choice] option.

+++ tiles/tiles.hpp
@@ -15,7 +15,7 @@ public:
 	typedef unsigned int Tile;
 	typedef unsigned int Pos;
 
-	typedef int Cost;
+	typedef float Cost;
 	typedef int Oper;
 	enum {
 		Nop = -1,
